+++ src/main/java/kitchenpos/product/presentation/dto/ProductCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.product.presentation.dto;
 
 import java.math.BigDecimal;
