+++ src/test/java/kitchenpos/menu/presentation/MenuRestControllerTest.java
@@ -3,9 +3,9 @@ package kitchenpos.menu.presentation;
 import io.restassured.RestAssured;
 import kitchenpos.common.controller.ControllerTest;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
-import kitchenpos.domain.repository.ProductRepository;
+import kitchenpos.product.domain.repository.ProductRepository;
 import kitchenpos.menu.presentation.dto.MenuCreateRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
